re-styled splash page
